Start with a fresh pyproject.toml file
This is, in effect, the start of a reboot of this repo. But rather than nuke
the whole thing and start 100% from scratch, I'm taking it one step at a
time. The first thing to do is to start afresh with the pyproject.toml file.
By the looks of things this started out as a dalliance with hatch(ling); I
don't really know that, and we tried but failed to get that going fine with
Textual actual, so here I go with what we use for Textual actual: poetry.

The revamp starts here.