Report the correct print location in the console
Fixes https://github.com/Textualize/textual/issues/3237

I'm undecided about this. It doesn't feel quite right to special-case this,
but on the other hand `print` is a form of logging that seems to work
differently in Textual vs the normal logging facility (which seems to do its
own dance to handle the call location, before it makes it to devtools).

This does correct the issue, this does ensure that `print` reports the
location the same as any `self.log` call. So this is a commit made to
solicit more detail on the design decisions in the core of Textual.